+++ src/index.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
+import { sanitizeProgressStorage } from './utils/storageSanitizer';
 
 const registerServiceWorker = () => {
   if (
@@ -56,6 +57,7 @@ const dismissStaticSplash = () => {
 
 window.dismissStaticSplash = dismissStaticSplash;
 
+sanitizeProgressStorage();
 registerServiceWorker();
 
 const root = ReactDOM.createRoot(document.getElementById('root'));